app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Ayron Felipe <ayron41@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2"
psycopg2-binary = "^2.8.6"
djangorestframework = "^3.12.4"
Markdown = "^3.3.4"
django-filter = "^2.4.0"
requests = "^2.25.1"
Shapely = "^1.7.1"
PyYAML = "^5.4.1"
uritemplate = "^3.0.1"
drf-yasg = "^1.20.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^20.8b1"
flake8 = "^3.9.1"
pytest-django = "^4.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
